refaktorisert ut noe kode 400linjer regelen
