+++ src/sider/aarsavregning/stegKomponenter/vurderingAarsavregning/komponenter/utils.ts
@@ -10,6 +10,9 @@ import {
   AarsavregningResponse,
   erBrukerSkattepliktigIHelePerioden,
 } from "../../../../../services/modules/aarsavregning/aarsavregning";
+import { Medlemskapsperiode } from "../../../../../services/modules/medlemavfolketrygden/medlemskapsperioder";
+import MKV from "../../../../../melosyskodeverk";
+import { sorterEtterISOFomDato } from "../../../../../utils/dato";
 
 const mapFeilmelding = (error: any) => {
   const feilmelding = "Finner ikke trygdeavgiftssats. Melosys har ikke satser for årene før 2014.";
@@ -64,3 +67,31 @@ export function beregnTrygdeavgiftsperioder(
     })
     .catch((error) => setBeregningError(mapFeilmelding(error)));
 }
+
+export const lagInnvilgetMedlemskapsPeriode = (medlemskapsperioder?: Medlemskapsperiode[]) => {
+  if (medlemskapsperioder && !Utils._isEmpty(medlemskapsperioder)) {
+    const sorterteInnvilgedePerioder = [...medlemskapsperioder]
+      .filter((periode) => periode.innvilgelsesResultat === MKV.Koder.innvilgelsesResultat.INNVILGET)
+      .sort(sorterEtterISOFomDato);
+    return {
+      fom: sorterteInnvilgedePerioder[0].fomDato,
+      tom: sorterteInnvilgedePerioder[sorterteInnvilgedePerioder.length - 1].tomDato,
+    };
+  }
+  return {
+    tom: undefined,
+    fom: undefined,
+  };
+};
+
+export const oppdaterNyttTotalbeloep = async (behandlingID: number, aarsavregningID: number, totalAvgift?: number) => {
+  return Api.Aarsavregning.oppdaterTotalBelop(
+    behandlingID,
+    {
+      avregning: {
+        nyttTotalbeloep: totalAvgift,
+      },
+    },
+    aarsavregningID,
+  );
+};
